move profile bottom sheet into itsown component
